backend/requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
PyMySQL
Werkzeug 
flask-cors
python-dotenv
gunicorn